laboratory2

+++ src/laboratory2/CalCul.java
@@ -44,11 +44,11 @@ public class CalCul {
         return num1 / num2;
     }
 
-    public static long division(long num1, long num2) {
-        return num1 / num2;
+    public static double division(long num1, long num2) {
+    return (num1 / num2);
     }
 
-    public static int division(int num1, int num2) {
+    public static double division(int num1, int num2) {
         return num1 / num2;
     }
 }
